core/meta: Move unique_variant stuff in meta
- It was cluttering the Result class and seems reusable
